requirements.txt
Flask
pymongo
python-docx
